Add docker-compose.yaml
Now a user should be able to start docker images for the postgres database, the
blockchain importer, and the backend all in with a single

  docker-compose up

The database will be initialized with the correct db schema and everything waits
for the database to come up.

The `backend` and `cardano-sl` submodules are pushed to new commits which are
required for the backend to work.